requirements.txt
# Telegram Bot Framework (asyncio-based)
python-telegram-bot[job-queue]>=20.7

# Groq AI Client (async compatible)
groq>=0.4.0

# Async SQLite support
aiosqlite>=0.19.0

# Async HTTP client (required by PTB and groq)
httpx>=0.25.0

# Type hints backport for Python 3.10
typing-extensions>=4.8.0

# Environment variable management
python-dotenv>=1.0.0

# Discord Bot
discord.py>=2.3.0

# Flask web server
flask>=3.0.0
